Ch06: Product flow stage 2 - #3 this is working on WildFly 8.1 for data population again.

+++ ch06/jsf-product-flow-stage2/src/main/java/uk/co/xenonique/digital/product/boundary/CampaignService.java
@@ -42,7 +42,7 @@ public class CampaignService {
         return query.getResultList();
     }
 
-    public List<Campaign> findById(Integer id) {
+    public List<Campaign> findById(Long id) {
         Query query = entityManager.createNamedQuery(
                 "Campaign.findById").setParameter("id", id);
         return query.getResultList();
