+++ ch06/jsf-product-flow-stage2/src/main/java/uk/co/xenonique/digital/product/boundary/UserProfileService.java
@@ -44,7 +44,7 @@ public class UserProfileService {
         return query.getResultList();
     }
 
-    public List<UserProfile> findById(Integer id) {
+    public List<UserProfile> findById(Long id) {
         Query query = entityManager.createNamedQuery(
             "UserProfile.findById").setParameter("id", id);
         return query.getResultList();
