+++ ch06/jsf-product-flow-stage2/src/main/java/uk/co/xenonique/digital/product/boundary/UserRoleService.java
@@ -45,7 +45,7 @@ public class UserRoleService {
         return query.getResultList();
     }
 
-    public List<UserRole> findById(Integer id) {
+    public List<UserRole> findById(Long id) {
         Query query = entityManager.createNamedQuery(
             "UserRole.findById").setParameter("id", id);
         return query.getResultList();
